ref(chunks): Rename `upload` module to `options`
I am planning to create a new `upload` module to contain a function which actually performs chunked uploading on generic objects. The `ChunkOptions` struct should stay in a separate module, hence moving it to the `options` module.
