+++ src/utils/chunks/mod.rs
@@ -5,11 +5,11 @@
 //!
 //! See `BatchedSliceExt::batches` for more information.
 
+mod options;
 mod types;
-mod upload;
 
+pub use options::ChunkOptions;
 pub use types::{Assemblable, Chunked, MissingObjectsInfo};
-pub use upload::ChunkOptions;
 
 use std::sync::Arc;
 use std::time::Duration;
